Add/use src-runtime/createTypeFromMapping.js
